+++ android/src/main/kotlin/com/flutter/logs/plogs/flutter_logs/LogsHelper.kt
@@ -1,11 +1,6 @@
 package com.flutter.logs.plogs.flutter_logs
 
-import android.Manifest
 import android.content.Context
-import android.content.pm.PackageManager
-import android.os.Environment
-import android.util.Log
-import androidx.core.content.ContextCompat
 import com.blackbox.plog.elk.PLogMetaInfoProvider
 import com.blackbox.plog.elk.models.fields.MetaInfo
 import com.blackbox.plog.mqtt.PLogMQTTProvider
@@ -45,76 +40,36 @@ object LogsHelper {
                     singleLogFileSize: Int?,
                     enabled: Boolean?) {
 
-        /*if (!permissionsGranted(context)) {
-            Log.e(TAG, "setUpLogger: Unable to setup logs. Permissions not granted.")
-            return
-        }*/
+        val config = LogsConfig(
+                logLevelsEnabled = logLevelsEnabled,
+                logTypesEnabled = logTypesEnabled,
+                logsRetentionPeriodInDays = logsRetentionPeriodInDays ?: 7,
+                zipsRetentionPeriodInDays = zipsRetentionPeriodInDays ?: 7,
+                autoDeleteZipOnExport = autoDeleteZipOnExport ?: false,
+                autoClearLogs = autoClearLogs ?: false,
+                autoExportErrors = autoExportErrors ?: false,
+                encryptionEnabled = encryptionEnabled ?: false,
+                encryptionKey = encryptionKey ?: "",
+                directoryStructure = getDirectoryStructure(directoryStructure),
+                logSystemCrashes = logSystemCrashes ?: false,
+                isDebuggable = isDebuggable ?: false,
+                debugFileOperations = debugFileOperations ?: false,
+                attachTimeStamp = attachTimeStamp ?: false,
+                attachNoOfFiles = attachNoOfFiles ?: false,
+                timeStampFormat = getTimeStampFormat(timeStampFormat),
+                logFileExtension = getLogFileExtension(logFileExtension),
+                zipFilesOnly = zipFilesOnly ?: false,
+                savePath = File(context.getExternalFilesDir(null), savePath).path,
+                zipFileName = zipFileName ?: "",
+                exportPath = File(context.getExternalFilesDir(null), savePath + File.separator + exportPath).path,
+                singleLogFileSize = singleLogFileSize ?: 1,
+                enableLogsWriteToFile = enabled ?: true
+        )
 
-        if (permissionsGranted(context)) {
-            createDir(savePath + File.separator + "Logs")
-
-            val config = LogsConfig(
-                    logLevelsEnabled = logLevelsEnabled,
-                    logTypesEnabled = logTypesEnabled,
-                    logsRetentionPeriodInDays = logsRetentionPeriodInDays ?: 7,
-                    zipsRetentionPeriodInDays = zipsRetentionPeriodInDays ?: 7,
-                    autoDeleteZipOnExport = autoDeleteZipOnExport ?: false,
-                    autoClearLogs = autoClearLogs ?: false,
-                    autoExportErrors = autoExportErrors ?: false,
-                    encryptionEnabled = encryptionEnabled ?: false,
-                    encryptionKey = encryptionKey ?: "",
-                    directoryStructure = getDirectoryStructure(directoryStructure),
-                    logSystemCrashes = logSystemCrashes ?: false,
-                    isDebuggable = isDebuggable ?: false,
-                    debugFileOperations = debugFileOperations ?: false,
-                    attachTimeStamp = attachTimeStamp ?: false,
-                    attachNoOfFiles = attachNoOfFiles ?: false,
-                    timeStampFormat = getTimeStampFormat(timeStampFormat),
-                    logFileExtension = getLogFileExtension(logFileExtension),
-                    zipFilesOnly = zipFilesOnly ?: false,
-                    savePath = createDir(savePath),
-                    zipFileName = zipFileName ?: "",
-                    exportPath = createDir(exportPath),
-                    singleLogFileSize = singleLogFileSize ?: 1,
-                    enabled = enabled ?: true
-            )
-
-            PLog.applyConfigurations(config, saveToFile = true, context = context)
-        } else {
-            val config = LogsConfig(
-                    logLevelsEnabled = logLevelsEnabled,
-                    logTypesEnabled = logTypesEnabled,
-                    logsRetentionPeriodInDays = logsRetentionPeriodInDays ?: 7,
-                    zipsRetentionPeriodInDays = zipsRetentionPeriodInDays ?: 7,
-                    autoDeleteZipOnExport = autoDeleteZipOnExport ?: false,
-                    autoClearLogs = autoClearLogs ?: false,
-                    autoExportErrors = autoExportErrors ?: false,
-                    encryptionEnabled = encryptionEnabled ?: false,
-                    encryptionKey = encryptionKey ?: "",
-                    directoryStructure = getDirectoryStructure(directoryStructure),
-                    logSystemCrashes = logSystemCrashes ?: false,
-                    isDebuggable = isDebuggable ?: false,
-                    debugFileOperations = debugFileOperations ?: false,
-                    attachTimeStamp = attachTimeStamp ?: false,
-                    attachNoOfFiles = attachNoOfFiles ?: false,
-                    timeStampFormat = getTimeStampFormat(timeStampFormat),
-                    logFileExtension = getLogFileExtension(logFileExtension),
-                    zipFilesOnly = zipFilesOnly ?: false,
-                    zipFileName = zipFileName ?: "",
-                    singleLogFileSize = singleLogFileSize ?: 1,
-                    enabled = enabled ?: true
-            )
-
-            PLog.applyConfigurations(config, context = context)
-        }
+        PLog.applyConfigurations(config, context = context)
     }
 
     fun writeLogToFile(context: Context, type: String, data: String?, appendTimeStamp: Boolean) {
-
-        if (!permissionsGranted(context)) {
-            Log.e(TAG, "writeLogToFile: Unable to setup logs. Permissions not granted.")
-            return
-        }
 
         try {
             if (appendTimeStamp) {
@@ -130,11 +85,6 @@ object LogsHelper {
     }
 
     fun overWriteLogToFile(context: Context, type: String, data: String?, appendTimeStamp: Boolean) {
-
-        if (!permissionsGranted(context)) {
-            Log.e(TAG, "overWriteLogToFile: Unable to setup logs. Permissions not granted.")
-            return
-        }
 
         try {
             if (appendTimeStamp) {
@@ -171,11 +121,6 @@ object LogsHelper {
                          longitude: String?,
                          labels: String?
     ) {
-
-        /*if (!permissionsGranted(context)) {
-            Log.e(TAG, "setupForELKStack: Unable to setup logs. Permissions not granted.")
-            return
-        }*/
 
         PLogMetaInfoProvider.elkStackSupported = true
 
@@ -221,10 +166,6 @@ object LogsHelper {
             initialDelaySecondsForPublishing: Int?
 
     ) {
-        /*if (!permissionsGranted(context)) {
-            Log.e(TAG, "setMQTT: Unable to setup logs. Permissions not granted.")
-            return
-        }*/
 
         if (brokerUrl.isNotEmpty()) {
             PLogMQTTProvider.initMQTTClient(context,
@@ -241,18 +182,5 @@ object LogsHelper {
                             ?: 30L
             )
         }
-    }
-
-    private fun createDir(pathName: String?): String {
-        val file = File(Environment.getExternalStorageDirectory().toString() + File.separator + pathName)
-        return if (!file.exists()) {
-            val result = file.mkdirs()
-            file.path
-        } else file.path
-    }
-
-    fun permissionsGranted(context: Context): Boolean {
-        return ContextCompat.checkSelfPermission(context, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED
-                && ContextCompat.checkSelfPermission(context, Manifest.permission.READ_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED
     }
 }