+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.6
+
+* Updated PLogs library version
+* Removed Runtime permissions for storage
+
 ## 0.0.1
 
 * TODO: Describe initial release.